header.php modified with $debug = true
I added a $debug variable to the header file that, when true, will
display debugging links on the experimenter header. Since pages have to
be restricted to either participants or experimenters, ONLY
EXPERIMENTERS CAN USE DATAEDIT.PHP for the time being. If this is a
problem we can fix it with a temporary hack.