eExperiments.php add Experiment functional
There is a link to the function on the page. The interface doesn't do
error checking and is pretty rudimentary, but it works. Ethnicities are
using placeholder values.

RELATED BUG: The insert query, for whatever reason, defaults to
inserting the entry with expid = 1 and will increment by 1 each time the
query is run. As a result, on a fresh \i database.sql the query will
fail 4 times until it gets to expid = 4 as up until that point it is
trying to insert an experiment with an expid which is already used,
violating the primary key uniqueness constraint. I'm not sure why this
is happening -- my guess is that it has something to do with the sql
script and not the query as the query looks fine and the problem is not
affecting other tables under similar circumstances.