pEnroll.php extra error checking for validate fxn
Added a bit more error checking. The default value of "none" for
experimenters->requirements will throw an error as it's not valid JSON.
Still can't do much testing as the pEnroll page isn't functioning.

Also did some testing for eExperiments and the new ethnicities seem to
be working fine. The add function of eSessions appears to function fine.
eExperiments is still erroring out on the first few new experiment
submits because of some database issue. It tries to insert with expid =
1 first, then 2, 3, etc and these are taken, causing database errors.